Display more data

+++ app/src/main/java/com/revengeos/weather/WeatherResponse.java
@@ -64,9 +64,9 @@ class Main {
     @SerializedName("temp")
     public float temp;
     @SerializedName("humidity")
-    public float humidity;
+    public int humidity;
     @SerializedName("pressure")
-    public float pressure;
+    public int pressure;
     @SerializedName("temp_min")
     public float temp_min;
     @SerializedName("temp_max")
